+++ test/platform/test_platform_context.cpp
@@ -23,7 +23,7 @@ struct TestServiceScope
 
   TestServiceScope()
   {
-    alloc.Init();
+    (void)SetAllocator(&alloc);
     jobs.Init();
     profiler.Init();
     logger.Init();
@@ -31,7 +31,6 @@ struct TestServiceScope
     (void)events.Init();
 
     Services svc {
-        .Allocator = &alloc,
         .JobSystem = &jobs,
         .Profiler = &profiler,
         .Logger = &logger,
@@ -44,6 +43,7 @@ struct TestServiceScope
   ~TestServiceScope()
   {
     UninstallServices();
+    ResetAllocator();
   }
 };